fix: allow build to correctly extract version info (.git) (#111)
* chore: checkout with .git folder

* feat: make paths in tgz backwards compatible

* chore: update tar command to include file transformation

* chore: remove rtl_433 subproject

* chore: use ubuntu-latest

* chore: deno fmt

* chore: update .gitignore and add VS Code settings and extensions

This commit updates the .gitignore file to include the .vscode folder and its contents, except for the settings.json and extensions.json files. It also adds the settings.json and extensions.json files to the .vscode folder. The settings.json file enables the Deno extension and sets the default formatter to denoland.vscode-deno. The extensions.json file recommends the denoland.vscode-deno extension.